Rename old LocalvariablesNamesEnhancerJava7, which also generates parameter names fields. This class can be dropped when we drop Java 7 support.
Plain LocalvariablesNamesEnhancer will not do it anymore, because parameter names are available using standard API in Java 8.
Build of Play now requires Java 8, but will still run with Java 7.
