+++ .travis.yml
@@ -1,7 +1,7 @@
 sudo: required
 language: java
 jdk:
-  - openjdk8 
+  - openjdk8
 #before_install:
 #  - cat /etc/hosts # optionally check the content *before*
 #  - sudo hostname "$(hostname | cut -c1-63)"
